Add cache for CheckRuns
The Checks-API in Gerrit is designed in a way that each Gerrit client
will send requests to the CI directly to get the Run status.
That means on a change that is watched by multiple reviewers at a
time, it doesn't make sense to compute the response every time.

This change adds a Caffeine cache that will cache the CheckRuns object
for each patchset for 30s. This basically limits the update interval
to 30s.

The expiry time can be made configurable in a future change.

Change-Id: I6062b4e8fc7d83d5f8f8593aca265678639820a1