Improve REST API endpoint path
The way the REST API in Jenkins work, the originally used "gerrit"
path prefix would be unique to the class setting it. No other class
would be able to use it.

Thus, the class name is too restricting, since additional request
types might be added later.

The prefix was also then chosen to be too general, since it might
be used by other plugins.

Change-Id: Ia78c6fc9de857e56ce5fb79423f579c61df3c281