Avoid NullPointerException if Build can't be found
If retrieving a Build based on the returned hits from Lucene, a
NullPointerException would be thrown in case that the Job or Build
is not present anymore in Jenkins. This Exception was not very
informative.

Now an IllegalStateException will be thrown, detailing what Job
or Build was not present anymore.

Change-Id: I70ec5491a6cc36c8fb52ff35183e1208ecf54bfc